Added support for A8,YUY2 and UYVY surfaces in dfb_surface_dump().

+++ src/core/surfaces.c
@@ -66,6 +66,7 @@
 #include <core/system.h>
 #include <core/windows_internal.h>
 
+#include <gfx/convert.h>
 #include <gfx/util.h>
 
 
@@ -935,13 +936,16 @@ DFBResult dfb_surface_dump( CoreSurface *surface,
                             const char  *prefix )
 {
      DFBResult          ret;
-     int                num = -1;
-     int                fd_p, fd_g = -1, i, n;
+     int                num  = -1;
+     int                fd_p = -1;
+     int                fd_g = -1;
+     int                i, n;
      int                len = strlen(directory) + strlen(prefix) + 40;
      char               filename[len];
      char               head[30];
      void              *data;
      int                pitch;
+     bool               rgb   = false;
      bool               alpha = false;
 #ifdef USE_ZLIB
      gzFile             gz_p, gz_g = NULL;
@@ -956,6 +960,10 @@ DFBResult dfb_surface_dump( CoreSurface *surface,
 
      /* Check pixel format. */
      switch (surface->format) {
+          case DSPF_A8:
+               alpha = true;
+               break;
+               
           case DSPF_ARGB:
           case DSPF_ARGB1555:
           case DSPF_AiRGB:
@@ -966,6 +974,9 @@ DFBResult dfb_surface_dump( CoreSurface *surface,
           case DSPF_RGB16:
           case DSPF_RGB24:
           case DSPF_RGB32:
+          case DSPF_YUY2:
+          case DSPF_UYVY:
+               rgb   = true;
                break;
 
           default:
@@ -979,26 +990,47 @@ DFBResult dfb_surface_dump( CoreSurface *surface,
      if (ret)
           return ret;
 
-     /* Create a file with the lowest unused pixmap index. */
-     do {
-          snprintf( filename, len, "%s/%s_%04d.ppm%s", directory, prefix, ++num, gz_ext );
+     /* Find the lowest unused index. */
+     while (++num < 10000) {
+          snprintf( filename, len, "%s/%s_%04d.ppm%s",
+                    directory, prefix, num, gz_ext );
+          
+          if (access( filename, F_OK ) != 0) {
+               snprintf( filename, len, "%s/%s_%04d.pgm%s",
+                         directory, prefix, num, gz_ext );
 
-          errno = 0;
+               if (access( filename, F_OK ) != 0)
+                    break;
+          }
+     }
+
+     if (num == 10000) {
+          D_ERROR( "DirectFB/core/surfaces: "
+                   "couldn't find an unused index for surface dump!\n" );
+          dfb_surface_unlock( surface, true );
+          return DFB_FAILURE;
+     }
+     
+     /* Create a file with the found index. */
+     if (rgb) {
+          snprintf( filename, len, "%s/%s_%04d.ppm%s",
+                    directory, prefix, num, gz_ext );
 
           fd_p = open( filename, O_EXCL | O_CREAT | O_WRONLY, 0644 );
-          if (fd_p < 0 && errno != EEXIST) {
+          if (fd_p < 0) {
                D_PERROR("DirectFB/core/input: "
-                         "could not open %s!\n", filename);
+                        "could not open %s!\n", filename);
 
                dfb_surface_unlock( surface, true );
 
                return DFB_IO;
           }
-     } while (errno == EEXIST);
+     }
 
-     /* Create a graymap for the alpha channel using the same index. */
+     /* Create a graymap for the alpha channel using the found index. */
      if (alpha) {
-          snprintf( filename, len, "%s/%s_%04d.pgm%s", directory, prefix, num, gz_ext );
+          snprintf( filename, len, "%s/%s_%04d.pgm%s",
+                    directory, prefix, num, gz_ext );
 
           fd_g = open( filename, O_EXCL | O_CREAT | O_WRONLY, 0644 );
           if (fd_g < 0) {
@@ -1007,31 +1039,35 @@ DFBResult dfb_surface_dump( CoreSurface *surface,
 
                dfb_surface_unlock( surface, true );
 
-               close( fd_p );
-
-               snprintf( filename, len, "%s/%s_%04d.ppm%s",
-                         directory, prefix, num, gz_ext );
-               unlink( filename );
+               if (rgb) {
+                    close( fd_p );
+                    snprintf( filename, len, "%s/%s_%04d.ppm%s",
+                              directory, prefix, num, gz_ext );
+                    unlink( filename );
+               }
 
                return DFB_IO;
           }
      }
 
 #ifdef USE_ZLIB
-     gz_p = gzdopen( fd_p, "wb" );
+     if (rgb)
+          gz_p = gzdopen( fd_p, "wb" );
 
      if (alpha)
           gz_g = gzdopen( fd_g, "wb" );
 #endif
 
-     /* Write the pixmap header. */
-     snprintf( head, 30,
-               "P6\n%d %d\n255\n", surface->width, surface->height );
+     if (rgb) {
+          /* Write the pixmap header. */
+          snprintf( head, 30,
+                    "P6\n%d %d\n255\n", surface->width, surface->height );
 #ifdef USE_ZLIB
-     gzwrite( gz_p, head, strlen(head) );
+          gzwrite( gz_p, head, strlen(head) );
 #else
-     write( fd_p, head, strlen(head) );
+          write( fd_p, head, strlen(head) );
 #endif
+     }
 
      /* Write the graymap header. */
      if (alpha) {
@@ -1060,6 +1096,9 @@ DFBResult dfb_surface_dump( CoreSurface *surface,
           data32 = (__u32*) data8;
 
           switch (surface->format) {
+               case DSPF_A8:
+                    direct_memcpy( &buf_g[0], data8, surface->width );
+                    break;
                case DSPF_AiRGB:
                     for (n=0, n3=0; n<surface->width; n++, n3+=3) {
                          buf_p[n3+0] = (data32[n] & 0xFF0000) >> 16;
@@ -1108,16 +1147,43 @@ DFBResult dfb_surface_dump( CoreSurface *surface,
                          buf_p[n3+2] = (data32[n] & 0x0000FF);
                     }
                     break;
+               case DSPF_YUY2:
+                    for (n=0, n3=0; n<surface->width/2; n++, n3+=6) {
+                         register __u32 y0, cb, y1, cr;
+                         y0 = (data32[n] & 0x000000FF);
+                         cb = (data32[n] & 0x0000FF00) >>  8;
+                         y1 = (data32[n] & 0x00FF0000) >> 16;
+                         cr = (data32[n] & 0xFF000000) >> 24;
+                         YCBCR_TO_RGB( y0, cb, cr,
+                                       buf_p[n3+0], buf_p[n3+1], buf_p[n3+2] );
+                         YCBCR_TO_RGB( y1, cb, cr,
+                                       buf_p[n3+3], buf_p[n3+4], buf_p[n3+5] );
+                    }
+                    break;
+               case DSPF_UYVY: 
+                    for (n=0, n3=0; n<surface->width/2; n++, n3+=6) {
+                         register __u32 y0, cb, y1, cr;
+                         cb = (data32[n] & 0x000000FF);
+                         y0 = (data32[n] & 0x0000FF00) >>  8;
+                         cr = (data32[n] & 0x00FF0000) >> 16;
+                         y1 = (data32[n] & 0xFF000000) >> 24;
+                         YCBCR_TO_RGB( y0, cb, cr,
+                                       buf_p[n3+0], buf_p[n3+1], buf_p[n3+2] );
+                         YCBCR_TO_RGB( y1, cb, cr,
+                                       buf_p[n3+3], buf_p[n3+4], buf_p[n3+5] );
+                    }
+                    break;
                default:
                     D_BUG( "unexpected pixelformat" );
                     break;
           }
 
           /* Write color buffer to pixmap file. */
+          if (rgb)
 #ifdef USE_ZLIB
-          gzwrite( gz_p, buf_p, surface->width * 3 );
+               gzwrite( gz_p, buf_p, surface->width * 3 );
 #else
-          write( fd_p, buf_p, surface->width * 3 );
+               write( fd_p, buf_p, surface->width * 3 );
 #endif
 
           /* Write alpha buffer to graymap file. */
@@ -1133,14 +1199,16 @@ DFBResult dfb_surface_dump( CoreSurface *surface,
      dfb_surface_unlock( surface, true );
 
 #ifdef USE_ZLIB
-     gzclose( gz_p );
+     if (rgb)
+          gzclose( gz_p );
 
      if (alpha)
           gzclose( gz_g );
 #endif
 
      /* Close pixmap file. */
-     close( fd_p );
+     if (rgb)
+          close( fd_p );
 
      /* Close graymap file. */
      if (alpha)
